Add files via upload
Hospitals currently using a manual system for the management of Patient's information. The current system requires numerous paper forms. forms are often lost in transit between department. But with the help of this project we can overcome from this problem. All information managed in an efficiently so that an institution's resources may be effectively utilizes HMS.